display:  Silence integer narrowing warni^Werrors

+++ client/display/display.cc
@@ -440,8 +440,8 @@ int main( int argc, char** argv )
 		little_endian = false;
 	}
 	
-	short raster_size[ 2 ] = { desc.height, desc.width };
-	short window_size[ 2 ] = { desc.height, desc.width };
+	short raster_size[ 2 ] = { (short) desc.height, (short) desc.width };
+	short window_size[ 2 ] = { (short) desc.height, (short) desc.width };
 	
 	if ( x_numerator > 1  &&  x_numerator <= 16 )
 	{
